frontend/requirements.txt
flask==3.0.3
openai==1.12.0
python-dotenv==1.0.0 